requirements.txt
asgiref==3.10.0
colorama==0.4.6
Django==5.2.8
djangorestframework==3.16.1
iniconfig==2.3.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
